+++ src/components/BottomNavigation.tsx
@@ -3,25 +3,26 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { MaterialCommunityIcons } from 'react-native-vector-icons';
 
 import Home from '../screens/Home';
-import TodoList from '../screens/TodoList';
+import TodoScreen from '../screens/TodoList';
+import About from '../screens/About';
 
 const Tab = createBottomTabNavigator();
 
 const BottomTabs: FC = () => {
-
-
     return (
         <Tab.Navigator
             initialRouteName="Home"
             tabBarOptions={{
                 activeTintColor: '#e91e63'
+                
             }}
+            
         >
             <Tab.Screen
                 name="Home"
                 component={Home}
                 options={{
-                    tabBarLabel: 'Home asdf',
+                    tabBarLabel: 'Home',
                     tabBarIcon: ({color,size})=>(
                         <MaterialCommunityIcons name="home" color={color} size={size} />
                     )
@@ -29,14 +30,26 @@ const BottomTabs: FC = () => {
             />
             <Tab.Screen
                 name="TodoList"
-                component={TodoList}
+                component={TodoScreen}
                 options={{
                     tabBarLabel: 'TodoList',
                     tabBarIcon: ({color,size})=>(
-                        <MaterialCommunityIcons name="home" color={color} size={size} />
+                        <MaterialCommunityIcons name="format-list-bulleted" color={color} size={size} />
+                    )
+                }}
+            />
+            <Tab.Screen
+                name="About"
+                component={About}
+                options={{
+                    tabBarLabel: 'About',
+                    tabBarIcon: ({color,size})=>(
+                        <MaterialCommunityIcons name="information-outline" color={color} size={size} />
                     )
                 }}
             />
             </Tab.Navigator>
-    )
-}
+    );
+};
+
+export default BottomTabs;