+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - 2020-02-25
+## [1.0.0] - 2020-02-25
 
 ### Changes
 
@@ -13,4 +13,61 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
     - Initial app config
+
+
+## [1.0.1] - 2020-04-15
+
+### Changes
+
+    - 
+
+### Added
+
     - Bottom Navigation (nw)
+
+## [1.0.2] - 2020-04-20
+
+### Changes
+
+    - Bottom Navigation (working)
+
+### Added
+
+    - HomePage screen, todoList screen
+    
+
+
+## [1.1.0] - 2020-04-30
+
+### Changes
+
+    - Splashscreen
+
+### Added
+
+    - todoList reducer without deleting tasks
+    - About screen
+
+## [1.2.0] - 2020-05-10
+
+### Changes
+
+    - about, homepage, todolist styles
+
+### Added
+
+    - deleting tasks
+
+## [1.3.0] - 2020-05-25
+
+### Changes
+
+    - deleting tasks by longPress
+    - todoList elements styles
+
+### Added
+
+    - 
+
+    
+